feat: Complete documentation import path migration from src. to crackseg.
- Migrated all documentation files from 'from src.' to 'from crackseg.' imports
- Updated 110+ documentation files across guides, reports, tutorials, and API docs
- Fixed broken internal links and reference inconsistencies
- Applied comprehensive QA review with automated fixes
- Created backup protection for all modified files
- Updated code snippets and examples to use correct import paths
- Maintained historical accuracy in documentation reports
- Implemented automated verification and cleanup scripts

Subtask 6.12: Finalize and commit documentation updates

+++ tests/integration/data/test_data_pipeline.py
@@ -49,11 +49,13 @@ def test_data_pipeline_end_to_end() -> None:
     # Create dummy data
     n = 8
     images: list[NDArray[np.uint8]] = [
-        np.random.randint(0, 255, (32, 32, 3), dtype=np.uint8)
+        np.random.randint(0, 255, (32, 32, 3), dtype=np.uint8).astype(np.uint8)
         for _ in range(n)
     ]
     masks: list[NDArray[np.uint8]] = [
-        np.random.randint(0, 2, (32, 32), dtype=np.uint8) * 255
+        (np.random.randint(0, 2, (32, 32), dtype=np.uint8) * 255).astype(
+            np.uint8
+        )
         for _ in range(n)
     ]
     # Compose transforms
